test: refactor regression test for #25675 to use itBundled
